+++ dynts/style/__init__.py
@@ -1,0 +1,4 @@
+'''Script for styling SIRO web pages
+'''
+from .layout import *
+from .themes import *