+++ dynts/style/themes/siro.py
@@ -1,0 +1,7 @@
+from djpcms.style import CssTheme
+
+theme_name = 'siro'
+body = '#E8EDF0'
+dark = '#353432'
+light = '#DCDCDC'
+